Add plots with matplotlib
This overcomes limitations of Altair.
Before the recent release, we had a hacky boxplot,
which consisted of hand-crafted boxes and rules.
This basic version did not have outlier flyers etc.

After the upgrade, Altair ships with native support
for advanced boxplots, but the columns cannot be
sorted (apparently by design).